+++ dev/Project/src/DTOs/Workers.java
@@ -10,7 +10,7 @@ import java.util.*;
 public class Workers {
 
     private static Workers workers_instance = null;
-    private Map<Long, Worker> workers_map;
+    private Map<Integer, Worker> workers_map;
 
     private Workers() {
         workers_map = new HashMap<>();
@@ -54,14 +54,14 @@ public class Workers {
         return available_workers;
     }
 
-    public Worker getWorker(long worker_id)
+    public Worker getWorker(int worker_id)
     {
         if(!workers_map.containsKey(worker_id))
             return null;
         return workers_map.get(worker_id);
     }
 
-    public Map<Long,Worker> getAllWorkers() {
+    public Map<Integer,Worker> getAllWorkers() {
         return workers_map;
     }
 
@@ -87,7 +87,7 @@ public class Workers {
         return workers_string;
     }
 
-    public boolean removeWorker(long id)
+    public boolean removeWorker(int id)
     {
         return (workers_map.remove(id)!=null);
     }
